+++ apps/yapms/src/lib/components/sidebar/sections/savedmaps/SaveMap.svelte
@@ -15,6 +15,8 @@
 	async function createMap() {
 		submitting = true;
 		const form = new FormData();
+		const json = generateJson();
+		console.log(json);
 		const newMapData = new File([JSON.stringify(generateJson())], 'data.json', {
 			type: 'application/json'
 		});
